HELP-10397: Use default value for `type` in group pickup (#6100)
Previously `type` was not required; if not included the action would
essentially be skipped. The schema was then updated to require `type`
but existing feature code callflows still had empty `data` objects
with no `type` defined - thus still "not working".

Default the `type` to `extension` as that is the common `type` used by
SmartPBX-configured feature codes.