[master] KZOO-20 and 21: allow assign rejected portin number and used_by check after completed portin (#6089)
* fix used_by in various stage of port_in state and phone_numbers

Set correct `used_by` in number doc when port_in is completed.

An earlier bug allows `used_by` is not set when a port_in number is
in `rejected` state. Although this has been fix in this PR,
the ramification of earlier numbers without this fix still exists.
Therefore, this commit checks if `used_by` is not set properly when
the port_in transitions to `completed` state.

It also fixes another bug that created phone_number as a result of
port_in completed will preserve `used_by` app. That has been missing.

* apply all numbers app and simplify code

* add a check to check used_by list comprehension

* use knm_numbers:assign_to_app

* update specs

* use get_portin_number in cb_port_request:check_number_portability

* properly check account id of the port req

* fix spec

* handle not found portin number in split_port_request